Max visual distance tweak and fix warning

+++ src/Battlescape/TileEngine.cpp
@@ -440,7 +440,7 @@ bool TileEngine::visible(BattleUnit *currentUnit, Tile *tile)
 			{
 				visibleDistance += t->getSmoke() / 3;
 			}
-			if (visibleDistance > MAX_VOXEL_VIEW_DISTANCE)
+			if (visibleDistance > (size_t)MAX_VOXEL_VIEW_DISTANCE)
 			{
 				unitSeen = false;
 				break;
